chore: rename back to app_errors
