+++ packages/cli/src/lib/types/app_errors.ts
